refactor(backend): 예외 클래스를 exception 서브패키지로 정리
- 도메인/애플리케이션 패키지마다 예외 클래스가 하나씩 늘어나 파일이
  많아져서 exception/ 하위로 모음 (도메인 클래스 파일 수 정리)
- #17에서 만든 InvalidRoomCodeException, RoomNotFoundException도 포함
- GlobalExceptionHandler에 InvalidRoomNameException(400),
  RoomHostRequiredException(403) 처리 추가
