+++ backend/src/main/java/com/sssok/presentation/api/common/GlobalExceptionHandler.java
@@ -1,7 +1,9 @@
 package com.sssok.presentation.api.common;
 
-import com.sssok.application.room.RoomNotFoundException;
-import com.sssok.domain.room.InvalidRoomCodeException;
+import com.sssok.domain.room.exception.InvalidRoomNameException;
+import com.sssok.application.room.exception.RoomNotFoundException;
+import com.sssok.domain.room.exception.InvalidRoomCodeException;
+import com.sssok.domain.room.exception.RoomHostRequiredException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -20,5 +22,17 @@ public class GlobalExceptionHandler {
     public ResponseEntity<ErrorResponse> handleInvalidRoomCode(InvalidRoomCodeException e) {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST)
             .body(new ErrorResponse("INVALID_ROOM_CODE", e.getMessage()));
+    }
+
+    @ExceptionHandler(InvalidRoomNameException.class)
+    public ResponseEntity<ErrorResponse> handleInvalidRoomName(InvalidRoomNameException e) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+            .body(new ErrorResponse("INVALID_ROOM_NAME", e.getMessage()));
+    }
+
+    @ExceptionHandler(RoomHostRequiredException.class)
+    public ResponseEntity<ErrorResponse> handleRoomHostRequired(RoomHostRequiredException e) {
+        return ResponseEntity.status(HttpStatus.FORBIDDEN)
+            .body(new ErrorResponse("ROOM_HOST_REQUIRED", e.getMessage()));
     }
 }
